#16 Appを別クラスに切り分けた
